requirements.txt
Django==3.2.11
djangorestframework>=3.13.1
beautifulsoup4>=4.10.0
requests>=2.27.1
lxml>=4.7.1
gunicorn>=20.1.0